refactor: implement custom views data fetching
