+++ packages/application-components/package.json
@@ -31,6 +31,7 @@
     "@babel/runtime-corejs3": "^7.22.15",
     "@commercetools-frontend/actions-global": "22.8.4",
     "@commercetools-frontend/application-config": "22.8.4",
+    "@commercetools-frontend/application-shell": "22.8.4",
     "@commercetools-frontend/application-shell-connectors": "22.8.4",
     "@commercetools-frontend/assets": "22.8.4",
     "@commercetools-frontend/constants": "22.8.4",
