+++ packages/application-components/src/index.ts
@@ -44,7 +44,6 @@ export type { TPageContentFull } from './components/page-content-containers/page
 
 // Custom views
 export { default as CustomPanel } from './components/custom-views/custom-panel';
-export { default as CustomViewLoader } from './components/custom-views/custom-view-loader';
 
 // Utilities
 export { default as PortalsContainer } from './components/portals-container';
