+++ packages/application-shell/package.json
@@ -58,6 +58,7 @@
     "@commercetools-uikit/loading-spinner": "^16.7.3",
     "@commercetools-uikit/notifications": "^16.7.3",
     "@commercetools-uikit/primary-button": "^16.7.3",
+    "@commercetools-uikit/secondary-button": "^16.7.3",
     "@commercetools-uikit/select-input": "^16.7.3",
     "@commercetools-uikit/spacings": "^16.7.3",
     "@commercetools-uikit/text": "^16.7.3",
