+++ packages/application-shell/src/components/custom-view-loader/index.ts
