+++ playground/src/components/custom-views/custom-panel-demo.jsx
@@ -1,8 +1,6 @@
 import { useMemo, useState } from 'react';
-import {
-  CustomViewLoader,
-  useModalState,
-} from '@commercetools-frontend/application-components';
+import { useModalState } from '@commercetools-frontend/application-components';
+import { CustomViewLoader } from '@commercetools-frontend/application-shell';
 import SecondaryButton from '@commercetools-uikit/secondary-button';
 import Spacings from '@commercetools-uikit/spacings';
 import Text from '@commercetools-uikit/text';
